+++ app/src/test/java/org/blitzortung/android/alert/handler/AlertDataHandlerTest.kt
@@ -43,8 +43,8 @@ class AlertDataHandlerTest {
         strike = DefaultStrike(timestamp = now, longitude = 12.0, latitude = 41.0, altitude = 0, amplitude = 20.0f)
 
         strikeLocation = Location("")
-        strikeLocation.longitude = strike.longitude.toDouble()
-        strikeLocation.latitude = strike.latitude.toDouble()
+        strikeLocation.longitude = strike.longitude
+        strikeLocation.latitude = strike.latitude
 
         parameters = AlertParameters(
                 alarmInterval = 10 * 60 * 1000,
